@@ -1,6 +1,7 @@
 package it.unive.dais.legodroid.lib.comm;
 
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 /**
  * This class implements a command for the GenEV3 device.
