@@ -1,9 +1,9 @@
 package it.unive.dais.legodroid.lib.comm;
 
-import android.support.annotation.NonNull;
-
 import java.io.IOException;
 import java.util.concurrent.Future;
+
+import androidx.annotation.NonNull;
 
 /**
  * Classes implementing this interface represent active on-going connections between two devices, in a similar fashion to {@link Channel}.
